requirements.txt
# Основные зависимости для бэкенда
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
python-telegram-bot>=20.6
cryptography>=41.0.4
requests>=2.31.0
python-dotenv>=1.0.0

# Дополнительные зависимости
python-multipart>=0.0.6  # для загрузки файлов в FastAPI
httpx>=0.25.0  # HTTP клиент для асинхронных запросов
jinja2>=3.1.2  # для шаблонов (если используются)
aiofiles>=23.2.1  # для асинхронной работы с файлами

# Зависимости для деплоя
gunicorn>=21.2.0  # WSGI HTTP сервер для продакшена
psycopg2-binary>=2.9.9  # для работы с PostgreSQL 